[main, cmd] Put subcommands in subpackage
Separates each subcommand into a different file for easier navigation.
The subcommand package is called gincmd
